add pretty matchers for shouldNot containsInAnyOrder
